requirements.txt
autogluon==1.0.0
# matplotlib==3.8.2
# numpy==1.26.2
# pandas==2.1.3
# Pillow==10.1.0
# scikit_learn==1.3.2
# timm==0.9.12
# torch==2.0.1
# torchvision==0.15.2
# tqdm==4.65.2
split-folders==0.5.1
google-cloud-storage==2.14.0
# ray==2.9.0
redis==5.0.1
celery==5.3.6
python-dotenv==1.0.0
uvicorn==0.25.0
fastapi==0.108.0
pydantic==1.10.13

platformdirs~=3.11.0
Pillow~=10.1.0
torch~=2.0.1
Jinja2~=3.1.3
pandas~=2.1.4
numpy~=1.26.3
nptyping~=2.4.1
joblib~=1.3.2
scikit-learn~=1.4.0
xgboost~=2.0.3
requests~=2.31.0
tabpfn~=0.1.10
optuna~=3.5.0